+++ src/app/InfinityProvider.ts
@@ -1,21 +1,24 @@
 import { CSVParser, JSONParser, XMLParser, HTMLParser } from './parsers';
 import { InfinityQuery, InfinityQuerySources, InfinityQueryType } from '../types';
 import { Datasource } from './../datasource';
+import { getTemplateSrv } from '@grafana/runtime';
 
 export class InfinityProvider {
   constructor(private target: InfinityQuery, private datasource: Datasource) {}
   private async formatResults(res: any) {
+    const query = this.target;
+    query.root_selector = getTemplateSrv().replace(query.root_selector);
     switch (this.target.type) {
       case InfinityQueryType.HTML:
-        return new HTMLParser(res, this.target).getResults();
+        return new HTMLParser(res, query).getResults();
       case InfinityQueryType.JSON:
       case InfinityQueryType.GraphQL:
-        return new JSONParser(res, this.target).getResults();
+        return new JSONParser(res, query).getResults();
       case InfinityQueryType.XML:
-        let xmldata = await new XMLParser(res, this.target);
-        return xmldata.getResults();
+        let xmlData = await new XMLParser(res, query);
+        return xmlData.getResults();
       case InfinityQueryType.CSV:
-        return new CSVParser(res, this.target).getResults();
+        return new CSVParser(res, query).getResults();
       default:
         return undefined;
     }
@@ -32,26 +35,6 @@ export class InfinityProvider {
         });
     });
   }
-  // private getRequestObject(target: InfinityQuery, instanceSettings: InfinityInstanceSettings): BackendSrvRequest {
-  //   let requestObject: BackendSrvRequest = {
-  //     url: target.url,
-  //     method: target.url_options && target.url_options.method ? target.url_options.method : 'GET',
-  //   };
-  //   if (instanceSettings.jsonData.datasource_mode === DatasourceMode.Advanced) {
-  //     const instanceSettingsUrl = instanceSettings.url;
-  //     const urlPath = target.url;
-  //     requestObject.url = [instanceSettingsUrl, urlPath].join('/');
-  //   }
-  //   if (target.url_options && target.url_options.method === 'POST') {
-  //     requestObject.data = target.url_options.data || '';
-  //     if (target.type === InfinityQueryType.GraphQL) {
-  //       requestObject.data = JSON.stringify({
-  //         query: `${target.url_options.data}`,
-  //       });
-  //     }
-  //   }
-  //   return requestObject;
-  // }
   private fetchResults() {
     return new Promise((resolve, reject) => {
       this.proxyResults(this.target)