+++ src/editors/query/query.url.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { ChangeEvent, useState } from 'react';
 import { set } from 'lodash';
 import { QueryColumnsEditor } from './query.columns.editor';
 import { URLOptionsEditor } from './query.url.options';
@@ -12,16 +12,17 @@ interface ScrapperProps {
 }
 
 export const URLEditor: React.FC<ScrapperProps> = props => {
-  const { query, onChange, onRunQuery } = props;
-  const [state, setState] = useState<InfinityQuery>(props.query);
-  const onSave = () => {
-    onChange(state);
-    onRunQuery();
-  };
-  const updateState = (value: string, key: keyof InfinityQuery) => {
-    let newQuery = { ...query };
-    set(newQuery, key, value);
-    setState(newQuery);
+  const [data, setData] = useState(props.query.data);
+  const onInputTextChange = (
+    event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
+    field: keyof InfinityQuery,
+    props: any,
+    splitIntoArray = false
+  ) => {
+    const { query, onChange } = props;
+    const value = splitIntoArray ? event.target.value.split(',') : event.target.value;
+    set(query, field, value);
+    onChange(query);
   };
   const LABEL_WIDTH = props.mode === EditorMode.Variable ? 10 : 8;
   return (
@@ -33,12 +34,12 @@ export const URLEditor: React.FC<ScrapperProps> = props => {
             <input
               type="text"
               className="gf-form-input min-width-30"
-              value={state.url}
+              value={props.query.url}
               placeholder="https://jsonplaceholder.typicode.com/todos"
-              onChange={e => updateState(e.currentTarget.value, `url`)}
-              onBlur={onSave}
+              onChange={e => onInputTextChange(e, `url`, props)}
+              onBlur={props.onRunQuery}
             ></input>
-            <URLOptionsEditor onChange={props.onChange} query={{ ...state }} />
+            <URLOptionsEditor onChange={props.onChange} query={props.query} />
           </div>
         </div>
       ) : (
@@ -48,32 +49,31 @@ export const URLEditor: React.FC<ScrapperProps> = props => {
             <textarea
               rows={5}
               className="gf-form-input min-width-30"
-              value={state.data}
+              value={data}
               placeholder=""
-              onBlur={onSave}
-              onChange={e => updateState(e.currentTarget.value, 'data')}
+              onBlur={e => onInputTextChange(e, `data`, props)}
+              onChange={e => setData(e.target.value)}
             ></textarea>
           </div>
         </div>
       )}
-      {['html', 'json', 'xml', 'graphql'].indexOf(state.type) > -1 ? (
+      {['html', 'json', 'xml', 'graphql'].indexOf(props.query.type) > -1 ? (
         <div className="gf-form-inline">
           <div className="gf-form">
             <label className={`gf-form-label query-keyword width-${LABEL_WIDTH}`}>Rows / Root</label>
             <input
               type="text"
               className="gf-form-input min-width-30"
-              value={state.root_selector}
+              value={props.query.root_selector}
               placeholder=""
-              onChange={e => updateState(e.currentTarget.value, `root_selector`)}
-              onBlur={onSave}
+              onChange={e => onInputTextChange(e, `root_selector`, props)}
             ></input>
           </div>
         </div>
       ) : (
         <></>
       )}
-      <QueryColumnsEditor onChange={props.onChange} query={{ ...query }} mode={props.mode} />
+      <QueryColumnsEditor onChange={props.onChange} query={props.query} mode={props.mode} />
     </>
   );
 };