+++ CHANGELOG.md
@@ -2,15 +2,16 @@
 
 Change history of the project. All the feature updates, bug fixes, breaking changes will be documented here.
 
-## [ 0.6.0 ]
+## [ 0.6.1 ]
 
 - Plugin signed
 - JSONPath support for JSON root selector
-- Autogenerate columns for CSV, JSON
+- Auto generate columns for CSV, JSON
 - Support for Unix/epoch seconds format
 - Support for filtering rows
 - Variable Query support
 - Logo updated
+- Published to [grafana.com](https://grafana.com/grafana/plugins/yesoreyeram-infinity-datasource/)
 
 ## [ 0.5.0 ]